% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hxmer_generator.R
\name{hxmer_generator}
\alias{hxmer_generator}
\title{hxmer_generator}
\usage{
hxmer_generator(num_words = 96, h_dist = 3, show_df = TRUE)
}
\arguments{
\item{num_words}{number of words in the list. Default to 96.}

\item{h_dist}{hamming distance of each seq. Default to 3.}

\item{show_df}{an option to show the distance matrix. Default to TRUE.}
}
\value{
a List of Words and a distance matrix
}
\description{
This function allows the user to generate a list of DNA seq and a distance matrix.
}
\examples{
hxmer_generator(num_words = 5, h_dist = 3, show_df = TRUE)
}
